requirements.txt
numpy==1.18.4
pandas==1.0.3
pytorch-ignite==0.3.0
sacred==0.8.1
scikit-learn==0.22.2.post1
sktime==0.3.1
torchdiffeq==0.1.0
tqdm==4.46.0
tensorboard==2.2.1
tensorboardX==2.1